We now cast the integers to int_least16_t to prevent warnings

+++ src/support/LCD.cpp
@@ -16,11 +16,11 @@ LCD::LCD()
  * Use the default foreground and backgroudnd colors.
  *
  * @param  char * text
- * @param  int_least16_t x
- * @param  int_least16_t y
+ * @param  int x
+ * @param  int y
  * @return void
  */
-void LCD::write(const char * text, int_least16_t x, int_least16_t y)
+void LCD::write(const char *text, int x, int y)
 {
-  this->drawText(x, y, text, RGB(255, 255, 255), RGB(24, 25, 30), 1);
+  this->drawText((int_least16_t) x, (int_least16_t) y, text, RGB(255, 255, 255), RGB(24, 25, 30), 1);
 }
